refactor: ArticleCard 컴포넌트 두가지 아티클 타입 받을 수 있게 수정(일반 아티클, 프로젝트 상세 아티클)
